replace fileUtility from front to back
