+++ admin/src/utils/utility.js
@@ -13,10 +13,15 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
+import byteSize from 'byte-size';
 
-export const EDIT_FORMATS = ['.docx', '.pptx', '.xlsx'];
+module.exports = {
 
-export const DOCUMENT_EXTS = ['.doc', '.docx', '.docm', '.dot', '.dotx', '.dotm', '.odt', '.fodt', '.ott', '.rtf', '.txt', '.html', '.htm', '.mht', '.xml', '.pdf', '.djvu', '.fb2', '.epub', '.xps', '.oxps'];
-export const SPREADSHEET_EXTS = ['.xls', '.xlsx', '.xlsm', '.xlt', '.xltx', '.xltm', '.ods', '.fods', '.ots', '.csv'];
-export const PRESENTATION_EXTS = ['.pps', '.ppsx', '.ppsm', '.ppt', '.pptx', '.pptm', '.pot', '.potx', '.potm', '.odp', '.fodp', '.otp'];
-
+  formatBytes: (receivedBytes, decimals = 0) => {
+    const {value, unit} = byteSize(receivedBytes * 1000, {precision: decimals});
+    if (!unit) {
+      return '0B';
+    }
+    return `${value}${unit.toUpperCase()}`;
+  }
+}
