+++ server/utils/index.js
@@ -15,10 +15,28 @@
 */
 'use strict';
 
+const EDIT_FORMATS = ['.docx', '.pptx', '.xlsx'];
+
+const DOCUMENT_EXTS = ['.doc', '.docx', '.docm', '.dot', '.dotx', '.dotm', '.odt', '.fodt', '.ott', '.rtf', '.txt', '.html', '.htm', '.mht', '.xml', '.pdf', '.djvu', '.fb2', '.epub', '.xps', '.oxps'];
+const SPREADSHEET_EXTS = ['.xls', '.xlsx', '.xlsm', '.xlt', '.xltx', '.xltm', '.ods', '.fods', '.ots', '.csv'];
+const PRESENTATION_EXTS = ['.pps', '.ppsx', '.ppsm', '.ppt', '.pptx', '.pptm', '.pot', '.potx', '.potm', '.odp', '.fodp', '.otp'];
+
 const getService = name => {
   return strapi.plugin('onlyoffice').service(name);
 };
 
 module.exports = {
   getService,
+
+  isFileEditable: (ext) => {
+    return EDIT_FORMATS.includes(ext);
+  },
+
+  isFileOpenable: (ext) => {
+    return DOCUMENT_EXTS.concat(SPREADSHEET_EXTS).concat(PRESENTATION_EXTS).includes(ext);
+  },
+
+  getFileType: (ext) => {
+    return DOCUMENT_EXTS.includes(ext) ? 'word' : SPREADSHEET_EXTS.includes(ext) ? 'cell' : PRESENTATION_EXTS.includes(ext) ? 'slide' : null;
+  },
 };
